Miscellaneous changes and additions

+++ shared/src/ritzow/sandbox/world/entity/Entity.java
@@ -52,7 +52,7 @@ public abstract class Entity implements Transportable {
 		/* optional implementation */
 		throw new UnsupportedOperationException(this.getClass().getTypeName() + " does not implement onCollision for entities");
 	}
-	public void onCollision(World world, Block block, float blockX, float blockY, float time) {/* optional implementation */}
+	public void onCollision(World world, Block block, int blockX, int blockY, float time) {/* optional implementation */}
 
 	/** @return true if the entity should be removed from the world **/
 	public abstract boolean getShouldDelete();
